+++ include/wifi/connection.h
@@ -12,5 +12,6 @@ void initiateHandshake();
 
 void acceptData();
 void sendPacket(JsonDocument& packet);
+void createAndSendPacket(uint8_t priority, std::string message, char* logMessage);
 
 #endif
