+++ include/wifi/packet.h
@@ -8,7 +8,4 @@ void handlePacket(JsonDocument packet);
 std::string unint8ArrayToHexString(uint8_t* oldArray, int len);
 
 void constructPacket(JsonDocument& packet, std::string packetType);
-
-void createAndSendPacket(uint8_t priority, std::string message, char* logMessage)
-
 #endif
